added favicon and redireected to coming soon page
